Bug fix : delete a middle node of list
Problem : Because this is doubly link list for queue, so we need to
consider prevNode and nextNode link result.
Method : We let the slow pointer to pointer to head. When fast pointer
walk into the end, slow pointer is in the middle. We use slow store
prevNode->next info and then modify nextNode prev to prevNode. Next, we
change prevNode next to middleNode next (nextNode) and delete middle
node.