Implement q_remove_head and q_remove_tail
Method : I use <list.h> linux list api including list_first_entry(find
the first element in queue and list_empty(check queue is empty or not)
to simplify code efficiently. And return the string of the element to
be deleted to sp. Before doing this, it is necessary to check if sp is a
null pointer or buffer size cannot hold the string. Implement
q_remove_tail is almost same as q_remove_head except use list_last_entry
to get the tail element.