Implement q_delete_dup
Method : Using a variable to keep track of the list node currently
possibly containing duplicate string, compare it continuously with
the next node. Upon findgin nodes with identical string, delete the
node directly until encountering a node with a different string. Then,
by checking for a duplicate flag, remove the originally compared node.
Finally, update the variable to record the next node with a different
string for the next round of comparisons.