Refactor q_delete_dup
Change : In the previous approach, we introduced a flag variable to
control whether duplicate occurrences were detected. This time, we
check at the beginning whether the current node and the next node have
the same string. If they are different, we move on to the next node.
However, if they are the same, we need to delete the node with the
duplicate string first before removing the originally compared node.
This process allows us to proceed to the next roung of comparisons.