Refactor q_delete_mid and fix error
Modify : We utilize two pointer variables to point to the first and last nodes of the queue. When these variables meet (in the case of an odd number of nodes) or when they are adjacent (in the case of an even number of nodes), it indicates that we have reached the middle node(s). At this point, we remove the node(s) from the sequence and release the resources associated with the node(s) using the q_release_element function.