Implement q_sort
First, we need to implement q_merge_sorted and q_middle. The former method is to merge two sorted lists, and the latter is to find the middle node. Therefore, we iteratively find a middle node to partition the list. If the list has only one node, we can start the merging process. Ultimately, the list will be sorted.